+++ examples/sfs/share/share.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/stackitcloud/stackit-sdk-go/core/config"
 	"github.com/stackitcloud/stackit-sdk-go/core/utils"
 	"github.com/stackitcloud/stackit-sdk-go/services/sfs"
 	"github.com/stackitcloud/stackit-sdk-go/services/sfs/wait"
@@ -17,7 +16,7 @@ func main() {
 	region := "eu01"
 
 	// Create a new API client, that uses default authentication and configuration
-	sfsClient, err := sfs.NewAPIClient(config.WithRegion(region))
+	sfsClient, err := sfs.NewAPIClient()
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "[sfs API] Creating API client: %v\n", err)
 		os.Exit(1)